requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0               # async PostgreSQL driver
alembic==1.13.3               # migrations

# Rate limiting
slowapi==0.1.9

# Storage
azure-storage-blob==12.23.1   # Azure Blob (only used in prod)
aiofiles==24.1.0              # local async file writes

# Config
pydantic-settings==2.4.0

# Dev / test
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2
locust==2.32.3
bandit==1.8.0